+++ python/analytics/requirements.txt
@@ -1,133 +1,8 @@
-accelerate==1.3.0
-annotated-types==0.7.0
-anyio==4.8.0
-argon2-cffi==23.1.0
-argon2-cffi-bindings==21.2.0
-arrow==1.3.0
-asttokens==3.0.0
-async-lru==2.0.4
-attrs==25.1.0
-babel==2.16.0
 beautifulsoup4==4.12.3
-bleach==6.2.0
-certifi==2024.12.14
-cffi==1.17.1
-charset-normalizer==3.4.1
-comm==0.2.2
-debugpy==1.8.12
-decorator==5.1.1
-defusedxml==0.7.1
-exceptiongroup==1.2.2
-executing==2.2.0
-fastjsonschema==2.21.1
-filelock==3.17.0
-fqdn==1.5.1
-fsspec==2024.12.0
-h11==0.14.0
-httpcore==1.0.7
-httpx==0.28.1
-huggingface-hub==0.28.0
-idna==3.10
-ipykernel==6.29.5
-ipython==8.31.0
-ipywidgets==8.1.5
-isoduration==20.11.0
-jedi==0.19.2
-Jinja2==3.1.5
-json5==0.10.0
-jsonpointer==3.0.0
-jsonschema==4.23.0
-jsonschema-specifications==2024.10.1
-jupyter==1.1.1
-jupyter-console==6.6.3
-jupyter-events==0.11.0
-jupyter-lsp==2.2.5
-jupyter_client==8.6.3
-jupyter_core==5.7.2
-jupyter_server==2.15.0
-jupyter_server_terminals==0.5.3
-jupyterlab==4.3.5
-jupyterlab_pygments==0.3.0
-jupyterlab_server==2.27.3
-jupyterlab_widgets==3.0.13
-markdown-it-py==3.0.0
-MarkupSafe==3.0.2
-matplotlib-inline==0.1.7
-mdurl==0.1.2
-mistune==3.1.1
-mpmath==1.3.0
-nbclient==0.10.2
-nbconvert==7.16.6
-nbformat==5.10.4
-nest-asyncio==1.6.0
-networkx==3.4.2
-notebook==7.3.2
-notebook_shim==0.2.4
-numpy==2.2.2
-nvidia-cublas-cu12==12.4.5.8
-nvidia-cuda-cupti-cu12==12.4.127
-nvidia-cuda-nvrtc-cu12==12.4.127
-nvidia-cuda-runtime-cu12==12.4.127
-nvidia-cudnn-cu12==9.1.0.70
-nvidia-cufft-cu12==11.2.1.3
-nvidia-curand-cu12==10.3.5.147
-nvidia-cusolver-cu12==11.6.1.9
-nvidia-cusparse-cu12==12.3.1.170
-nvidia-cusparselt-cu12==0.6.2
-nvidia-nccl-cu12==2.21.5
-nvidia-nvjitlink-cu12==12.4.127
-nvidia-nvtx-cu12==12.4.127
 ollama==0.4.7
-overrides==7.7.0
-packaging==24.2
-pandocfilters==1.5.1
-parso==0.8.4
-pexpect==4.9.0
-platformdirs==4.3.6
-prometheus_client==0.21.1
-prompt_toolkit==3.0.50
-psutil==6.1.1
-ptyprocess==0.7.0
-pure_eval==0.2.3
-pycparser==2.22
-pydantic==2.10.6
-pydantic_core==2.27.2
-Pygments==2.19.1
+numpy==2.2.2
 PyPDF2==3.0.1
-python-dateutil==2.9.0.post0
-python-json-logger==3.2.1
-PyYAML==6.0.2
-pyzmq==26.2.0
-referencing==0.36.2
-regex==2024.11.6
+rich>=13.0.0
 requests==2.32.3
-rfc3339-validator==0.1.4
-rfc3986-validator==0.1.1
-rich==13.9.4
-rpds-py==0.22.3
-safetensors==0.5.2
-Send2Trash==1.8.3
-six==1.17.0
-sniffio==1.3.1
-soupsieve==2.6
-stack-data==0.6.3
-sympy==1.13.1
-terminado==0.18.1
-tinycss2==1.4.0
-tokenizers==0.21.0
-tomli==2.2.1
-torch==2.6.0
-tornado==6.4.2
 tqdm==4.67.1
-traitlets==5.14.3
-transformers==4.48.1
-triton==3.2.0
-types-python-dateutil==2.9.0.20241206
-typing_extensions==4.12.2
-uri-template==1.3.0
 urllib3==2.3.0
-wcwidth==0.2.13
-webcolors==24.11.1
-webencodings==0.5.1
-websocket-client==1.8.0
-widgetsnbextension==4.0.13
