+++ tutorials/agents-research/README.md
@@ -1,0 +1,8 @@
+Agents Research
+
+- course
+  - https://llmagents-learning.org/sp25
+  
+
+- Youtube Playlist
+  - https://www.youtube.com/playlist?list=PLS01nW3RtgopsNLeM936V4TNSsvvVglLc
